Allow deleting and duplicating a selection of rows
Before bbac655499d5f4d2334bb7068c47f120be2e9225 it was possible to delete
a set of selected rows. This makes it possible again by only selecting the
row if it is not already inside the selected rows. See issue #1283.

Additionally and for coherence, the "Duplicate record" from the context
menu is also made to apply to the list of selected rows. See issue #1090