Issue #1461: Binary cell dump not visible - with no scrollbars
QHexEdit widget was being set disabled instead of read-only, according to
comments, because there wasn't any setReadOnly in qhexedit. At least in
the current version, the method actually exists, so it is now used.

This allows the user to move the scroll and select text in the binary mode
of the 'Edit Database Cell' dock.