New shortcut key for copying cell value from above
Ctrl+' is used by spreadsheets for copying formula/value from the cell
above.

Our shortcut copies the value inside the inline editor when it is text.
That mimics the behaviour of LibreOffice Calc. For binary data that cannot
be edited inline, the BLOB is just copied from the above cell to the
current one.

See discussion in PR #1516.