Fix XML filter for import in Edit Database Cell dock and more.
The XML branch in the import was incorrectly programmed. Now it sets the
appropriate selected filter.

The list of filters is initialised once and the selected filter is passed
to the dialog using the specific parameter.

Chopped the last space in the image formats list for better style.

Case insensitivity applied to the suffix check as done in other parts of
the code.

See issue #1535.