Never mark the DB dirty if it is read-only. Issue #1514
Setting save points does not make sense for read-only databases, so
do never set a save-point while executing a user query if the database has
been opened in read-only mode.

If setSavePoint is ever called in read-only mode, a warning is printed and
the save-point is not set.

If ever a read-write query is executed, it would be rejected by SQLite so
there is no risk of immediately committing any change to the DB.