Foreign key combobox in table browser #614
This adds a combobox for the editor delegate when the associated column
has a foreign key constraint. In this way the user can choose the key value
based from a combo-box.

Known problems:
- When the editor delegate is activated the combobox switches to the first
value. It should preserve the current value, as intended in setEditorData.

-The encoding of the foreign table is the one from the current table, but
they might be different (we allow that, although it's unconventional).